pkp/pkp-lib#6351 Separate settings management for site and context levels
